// services/api/core/models/enum/EApiFilters.ts
export enum EApiFilters {
    EQUALS = 'equals',
    NOT_EQUALS = 'notEquals',
    IN = 'in',
    NOT_IN = 'notIn',
    CONTAINS = 'contains',
    DOES_NOT_CONTAIN = 'doesNotContain',
    GREATER_THAN = 'greaterThan',
    LESS_THAN = 'lessThan',
    GREATER_THAN_OR_EQUAL = 'greaterThanOrEqual',
    LESS_THAN_OR_EQUAL = 'lessThanOrEqual',
}